[addons][filesystem] separate "C" and "C++" parts of headers
By them becomes the "C" part placed in own header to see better what language
is where and to make easier for other languages.
