+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/Filesystem.h
@@ -9,263 +9,12 @@
 #pragma once
 
 #include "AddonBase.h"
-
-#include <stdint.h>
-#include <time.h>
+#include "c-api/filesystem.h"
 
 #ifdef __cplusplus
+
 #include <map>
 #include <vector>
-#endif /* __cplusplus */
-
-#ifdef _WIN32 // windows
-#ifndef _SSIZE_T_DEFINED
-typedef intptr_t ssize_t;
-#define _SSIZE_T_DEFINED
-#endif // !_SSIZE_T_DEFINED
-#endif
-
-#ifdef __cplusplus
-extern "C"
-{
-#endif /* __cplusplus */
-
-  //¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯
-  // "C" Definitions, structures and enumerators of filesystem
-  //{{{
-
-  //============================================================================
-  /// @defgroup cpp_kodi_vfs_Defs_OpenFileFlags enum OpenFileFlags
-  /// @ingroup cpp_kodi_vfs_Defs
-  /// @brief **Flags to define way how file becomes opened**\n
-  /// The values can be used together, e.g. <b>`file.Open("myfile", READ_TRUNCATED | READ_CHUNKED);`</b>
-  ///
-  /// Used on @ref kodi::vfs::CFile::OpenFile().
-  ///
-  //@{
-  typedef enum OpenFileFlags
-  {
-    /// @brief Indicate that caller can handle truncated reads, where function
-    /// returns before entire buffer has been filled.
-    READ_TRUNCATED = 0x01,
-
-    /// @brief Indicate that that caller support read in the minimum defined
-    /// chunk size, this disables internal cache then.
-    READ_CHUNKED = 0x02,
-
-    /// @brief Use cache to access this file.
-    READ_CACHED = 0x04,
-
-    /// @brief Open without caching. regardless to file type.
-    READ_NO_CACHE = 0x08,
-
-    /// @brief Calcuate bitrate for file while reading.
-    READ_BITRATE = 0x10,
-
-    /// @brief Indicate to the caller we will seek between multiple streams in
-    /// the file frequently.
-    READ_MULTI_STREAM = 0x20,
-
-    /// @brief indicate to the caller file is audio and/or video (and e.g. may
-    /// grow).
-    READ_AUDIO_VIDEO = 0x40,
-
-    /// @brief Indicate that caller will do write operations before reading.
-    READ_AFTER_WRITE = 0x80,
-
-    /// @brief Indicate that caller want to reopen a file if its already open.
-    READ_REOPEN = 0x100
-  } OpenFileFlags;
-  //@}
-  //----------------------------------------------------------------------------
-
-  //============================================================================
-  /// @defgroup cpp_kodi_vfs_Defs_CURLOptiontype enum CURLOptiontype
-  /// @ingroup cpp_kodi_vfs_Defs
-  /// @brief **CURL message types**\n
-  /// Used on kodi::vfs::CFile::CURLAddOption().
-  ///
-  //@{
-  typedef enum CURLOptiontype
-  {
-    /// @brief Set a general option.
-    ADDON_CURL_OPTION_OPTION,
-
-    /// @brief Set a protocol option.
-    ///
-    /// The following names for *ADDON_CURL_OPTION_PROTOCOL* are possible:
-    ///
-    /// | Option name                         | Description
-    /// |------------------------------------:|:--------------------------------
-    /// | <b>`accept-charset`</b>             | Set the "accept-charset" header
-    /// | <b>`acceptencoding or encoding`</b> | Set the "accept-encoding" header
-    /// | <b>`active-remote`</b>              | Set the "active-remote" header
-    /// | <b>`auth`</b>                       | Set the authentication method. Possible values: any, anysafe, digest, ntlm
-    /// | <b>`connection-timeout`</b>         | Set the connection timeout in seconds
-    /// | <b>`cookie`</b>                     | Set the "cookie" header
-    /// | <b>`customrequest`</b>              | Set a custom HTTP request like DELETE
-    /// | <b>`noshout`</b>                    | Set to true if kodi detects a stream as shoutcast by mistake.
-    /// | <b>`postdata`</b>                   | Set the post body (value needs to be base64 encoded). (Implicitly sets the request to POST)
-    /// | <b>`referer`</b>                    | Set the "referer" header
-    /// | <b>`user-agent`</b>                 | Set the "user-agent" header
-    /// | <b>`seekable`</b>                   | Set the stream seekable. 1: enable, 0: disable
-    /// | <b>`sslcipherlist`</b>              | Set list of accepted SSL ciphers.
-    ///
-    ADDON_CURL_OPTION_PROTOCOL,
-
-    /// @brief Set User and password
-    ADDON_CURL_OPTION_CREDENTIALS,
-
-    /// @brief Add a Header
-    ADDON_CURL_OPTION_HEADER
-  } CURLOptiontype;
-  //@}
-  //----------------------------------------------------------------------------
-
-  //============================================================================
-  /// @defgroup cpp_kodi_vfs_Defs_FilePropertyTypes enum FilePropertyTypes
-  /// @ingroup cpp_kodi_vfs_Defs
-  /// @brief **File property types**\n
-  /// Mostly to read internet sources.
-  ///
-  /// Used on kodi::vfs::CFile::GetPropertyValue() and kodi::vfs::CFile::GetPropertyValues().
-  ///
-  //@{
-  typedef enum FilePropertyTypes
-  {
-    /// @brief Get protocol response line.
-    ADDON_FILE_PROPERTY_RESPONSE_PROTOCOL,
-    /// @brief Get a response header.
-    ADDON_FILE_PROPERTY_RESPONSE_HEADER,
-    /// @brief Get file content type.
-    ADDON_FILE_PROPERTY_CONTENT_TYPE,
-    /// @brief Get file content charset.
-    ADDON_FILE_PROPERTY_CONTENT_CHARSET,
-    /// @brief Get file mime type.
-    ADDON_FILE_PROPERTY_MIME_TYPE,
-    /// @brief Get file effective URL (last one if redirected).
-    ADDON_FILE_PROPERTY_EFFECTIVE_URL
-  } FilePropertyTypes;
-  //@}
-  //----------------------------------------------------------------------------
-
-  //}}}
-
-  //¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯
-  // "C" Internal interface tables for intercommunications between addon and kodi
-  //{{{
-
-  struct STAT_STRUCTURE
-  {
-    /// ID of device containing file
-    uint32_t deviceId;
-    /// Total size, in bytes
-    uint64_t size;
-    /// Time of last access
-    time_t accessTime;
-    /// Time of last modification
-    time_t modificationTime;
-    /// Time of last status change
-    time_t statusTime;
-    /// The stat url is a directory
-    bool isDirectory;
-    /// The stat url is a symbolic link
-    bool isSymLink;
-  };
-
-  struct VFS_CACHE_STATUS_DATA
-  {
-    uint64_t forward;
-    unsigned int maxrate;
-    unsigned int currate;
-    bool lowspeed;
-  };
-
-  struct VFSProperty
-  {
-    char* name;
-    char* val;
-  };
-
-  struct VFSDirEntry
-  {
-    char* label; //!< item label
-    char* title; //!< item title
-    char* path; //!< item path
-    unsigned int num_props; //!< Number of properties attached to item
-    struct VFSProperty* properties; //!< Properties
-    time_t date_time; //!< file creation date & time
-    bool folder; //!< Item is a folder
-    uint64_t size; //!< Size of file represented by item
-  };
-
-  typedef struct AddonToKodiFuncTable_kodi_filesystem
-  {
-    bool (*can_open_directory)(void* kodiBase, const char* url);
-    bool (*create_directory)(void* kodiBase, const char* path);
-    bool (*remove_directory)(void* kodiBase, const char* path);
-    bool (*directory_exists)(void* kodiBase, const char* path);
-    bool (*get_directory)(void* kodiBase,
-                          const char* path,
-                          const char* mask,
-                          struct VFSDirEntry** items,
-                          unsigned int* num_items);
-    void (*free_directory)(void* kodiBase, struct VFSDirEntry* items, unsigned int num_items);
-
-    bool (*file_exists)(void* kodiBase, const char* filename, bool useCache);
-    bool (*stat_file)(void* kodiBase, const char* filename, struct STAT_STRUCTURE* buffer);
-    bool (*delete_file)(void* kodiBase, const char* filename);
-    bool (*rename_file)(void* kodiBase, const char* filename, const char* newFileName);
-    bool (*copy_file)(void* kodiBase, const char* filename, const char* dest);
-
-    char* (*get_file_md5)(void* kodiBase, const char* filename);
-    char* (*get_cache_thumb_name)(void* kodiBase, const char* filename);
-    char* (*make_legal_filename)(void* kodiBase, const char* filename);
-    char* (*make_legal_path)(void* kodiBase, const char* path);
-    char* (*translate_special_protocol)(void* kodiBase, const char* strSource);
-    bool (*is_internet_stream)(void* kodiBase, const char* path, bool strictCheck);
-    bool (*is_on_lan)(void* kodiBase, const char* path);
-    bool (*is_remote)(void* kodiBase, const char* path);
-    bool (*is_local)(void* kodiBase, const char* path);
-    bool (*is_url)(void* kodiBase, const char* path);
-
-    void* (*open_file)(void* kodiBase, const char* filename, unsigned int flags);
-    void* (*open_file_for_write)(void* kodiBase, const char* filename, bool overwrite);
-    ssize_t (*read_file)(void* kodiBase, void* file, void* ptr, size_t size);
-    bool (*read_file_string)(void* kodiBase, void* file, char* szLine, int iLineLength);
-    ssize_t (*write_file)(void* kodiBase, void* file, const void* ptr, size_t size);
-    void (*flush_file)(void* kodiBase, void* file);
-    int64_t (*seek_file)(void* kodiBase, void* file, int64_t position, int whence);
-    int (*truncate_file)(void* kodiBase, void* file, int64_t size);
-    int64_t (*get_file_position)(void* kodiBase, void* file);
-    int64_t (*get_file_length)(void* kodiBase, void* file);
-    double (*get_file_download_speed)(void* kodiBase, void* file);
-    void (*close_file)(void* kodiBase, void* file);
-    int (*get_file_chunk_size)(void* kodiBase, void* file);
-    bool (*io_control_get_seek_possible)(void* kodiBase, void* file);
-    bool (*io_control_get_cache_status)(void* kodiBase,
-                                        void* file,
-                                        struct VFS_CACHE_STATUS_DATA* status);
-    bool (*io_control_set_cache_rate)(void* kodiBase, void* file, unsigned int rate);
-    bool (*io_control_set_retry)(void* kodiBase, void* file, bool retry);
-    char** (*get_property_values)(
-        void* kodiBase, void* file, int type, const char* name, int* numValues);
-
-    void* (*curl_create)(void* kodiBase, const char* url);
-    bool (*curl_add_option)(
-        void* kodiBase, void* file, int type, const char* name, const char* value);
-    bool (*curl_open)(void* kodiBase, void* file, unsigned int flags);
-  } AddonToKodiFuncTable_kodi_filesystem;
-
-  //}}}
-
-#ifdef __cplusplus
-} /* extern "C" */
-#endif /* __cplusplus */
-
-//··············································································
-
-#ifdef __cplusplus
 
 namespace kodi
 {
